feat(chat): show group photo in the sidebar list and chat header
Extracted GroupAvatar to a shared component and use it in the conversation list
row, the thread header, and the members modal — so an uploaded group photo
appears everywhere the group is shown, not just inside the details modal.
